Command/requirements.txt
python-dotenv
slack_sdk
requests
boto3